Rename model_trainer.py to trainer.py
