+++ netspresso/trainer/__init__.py
@@ -1,4 +1,4 @@
 from netspresso.trainer.enums import Task
-from netspresso.trainer.model_trainer import Trainer
+from netspresso.trainer.trainer import Trainer
 
 __all__ = ["Trainer", "Task"]
